Slow stdin fixed
