+++ include/misc.h
@@ -6,6 +6,6 @@
 bool    load_opts(int argc, char *[]);
 long    memused(void);
 std::string  ptr2hex(std::string_view data);
-int     hex2bytes(const std::string&, char *);
+int     hex2bytes(std::string_view, char *);
 
 #endif // MISC_H
