+++ src/main.cpp
@@ -29,13 +29,13 @@ using namespace std;
 
 int     main(int argc, char *argv[]) {
   // TODO: local BUFFER = char *const ptr;
+  char BUFFER[MAX_BK_SIZE];
   bool (*bkloader)(char *, const uint32_t) = &stdin_bk;
 
     // 1. prepare
     // 1.1. handle options
     if (!load_opts(argc, argv))
       return 1;
-    // TODO: check options after all
     // 1.2. prepare bk info
     if (!OPTS.fromcin) {
       auto bk_qty = init_bkloader(OPTS.datdir, OPTS.locsfile);
@@ -50,7 +50,6 @@ int     main(int argc, char *argv[]) {
     if (!set_cache())
         return 3;
     // 1.4. last prestart
-    char *const BUFFER = new char[MAX_BK_SIZE];
     if (OPTS.verbose)
       __prn_head();
     start_time = time(nullptr);
